Refactor axios as hook and apply
