+++ src/hook/useUser.tsx
@@ -1,0 +1,14 @@
+import useApi from './useApi';
+
+function useUser() {
+  const { adminApi } = useApi();
+
+  const isLoggedIn = () => {
+    adminApi.get('/user');
+    return true;
+  };
+
+  return { isLoggedIn };
+}
+
+export default useUser;
