+++ src/page/Admin/AdminHome.tsx
@@ -1,19 +1,13 @@
-import React, { useEffect, useState } from 'react';
-import { adminApi } from '../../axios';
+import React from 'react';
+import useUser from '../../hook/useUser';
 
 function AdminHome() {
-  const [url, setUrl] = useState();
-
-  useEffect(() => {
-    adminApi.get('/user').then((res) => {
-      setUrl(res.data.accountUrl);
-    });
-  }, []);
+  const { isLoggedIn } = useUser();
 
   return (
     <div>
       <div>AdminHome 여기가 헤더가 되어야 할듯?</div>
-      <button onClick={() => window.open(`${url}&amount=5000`)}>test</button>
+      isLoggedIn: {isLoggedIn() ? 'true' : 'false'}
     </div>
   );
 }
